complete comment and edit rating

+++ src/main/java/vn/ute/smartphoneshop/repository/RatingRepository.java
@@ -8,6 +8,5 @@ import java.util.List;
 
 @Repository
 public interface RatingRepository extends JpaRepository<RatingEntity, Integer> {
-    List<RatingEntity> findByProduct_ProductId(Long productId);
-
+    List<RatingEntity> findByProduct_ProductId(int productId);
 }
